add def_static_kernel_args

+++ include/pybind/call/function_call.hpp
@@ -60,8 +60,8 @@ namespace pybind
             if( arg_size != fn_arity )
             {
                 pybind::throw_exception( "invalid function call args is not equal %d != %d (%s)"
-                    , (uint32_t)arg_size
-                    , (uint32_t)fn_arity
+                    , arg_size
+                    , fn_arity
                     , _kernel->object_repr( _arg ).c_str()
                 );
