+++ include/pybind/call/function_kernel_args_call.hpp
@@ -61,8 +61,8 @@ namespace pybind
             if( arg_size + 2 < fn_arity )
             {
                 pybind::throw_exception( "invalid args function call args is not equal %u < %u (%s)"
-                    , (uint32_t)arg_size + 2
-                    , (uint32_t)fn_arity
+                    , arg_size + 2
+                    , fn_arity
                     , _kernel->object_repr_type( _arg ).c_str()
                 );
