+++ include/pybind/call/function_proxy_call.hpp
@@ -62,8 +62,8 @@ namespace pybind
             if( arg_size + 1 != fn_arity )
             {
                 pybind::throw_exception( "invalid proxy function call args is not equal %d != %d (%s)"
-                    , (uint32_t)arg_size + 1
-                    , (uint32_t)fn_arity
+                    , arg_size + 1
+                    , fn_arity
                     , _kernel->object_repr( _arg ).c_str()
                 );
